feat: update imports, configs
