+++ python/beeai_framework/adapters/acp/serve/server.py
@@ -12,19 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections.abc import AsyncGenerator
-from datetime import UTC, datetime
+import asyncio
+import os
+from collections.abc import AsyncGenerator, Awaitable, Callable
+from datetime import UTC, datetime, timedelta
 from typing import Any, Generic, Self
 
 import acp_sdk.models as acp_models
 import acp_sdk.server.context as acp_context
 import acp_sdk.server.server as acp_server
 import acp_sdk.server.types as acp_types
+import uvicorn
 from acp_sdk import AnyModel, Author, Capability, Contributor, Dependency, Link
+from pydantic import BaseModel
 from typing_extensions import TypedDict, TypeVar, Unpack, override
 
 from beeai_framework.adapters.acp.serve._utils import acp_msg_to_framework_msg
-from beeai_framework.adapters.acp.serve.agent import ACPAgent, ACPServerConfig
+from beeai_framework.adapters.acp.serve.agent import ACPServerAgent
 from beeai_framework.agents import AnyAgent
 from beeai_framework.agents.react.agent import ReActAgent
 from beeai_framework.agents.react.events import ReActAgentUpdateEvent
@@ -64,8 +68,8 @@ class ACPServerMetadata(TypedDict, total=False):
     extra: dict[str, Any]
 
 
-class ACPServer(Generic[AnyAgentLike], Server[AnyAgentLike, ACPAgent, ACPServerConfig]):
-    def __init__(self, *, config: ModelLike[ACPServerConfig] | None = None) -> None:
+class ACPServer(Generic[AnyAgentLike], Server[AnyAgentLike, ACPServerAgent, "ACPServerConfig"]):
+    def __init__(self, *, config: ModelLike["ACPServerConfig"] | None = None) -> None:
         super().__init__(config=to_model(ACPServerConfig, config or {"self_registration": False}))
         self._metadata_by_agent: dict[AnyAgentLike, ACPServerMetadata] = {}
         self._server = acp_server.Server()
@@ -110,7 +114,7 @@ def to_acp_agent_metadata(metadata: ACPServerMetadata) -> acp_models.Metadata:
     return model
 
 
-def _react_agent_factory(agent: ReActAgent, *, metadata: ACPServerMetadata | None = None) -> ACPAgent:
+def _react_agent_factory(agent: ReActAgent, *, metadata: ACPServerMetadata | None = None) -> ACPServerAgent:
     if metadata is None:
         metadata = {}
 
@@ -135,7 +139,7 @@ def _react_agent_factory(agent: ReActAgent, *, metadata: ACPServerMetadata | Non
                         case "final_answer":
                             yield acp_models.MessagePart(content=update, role="assistant")  # type: ignore[call-arg]
 
-    return ACPAgent(
+    return ACPServerAgent(
         fn=run,
         name=metadata.get("name", agent.meta.name),
         description=metadata.get("description", agent.meta.description),
@@ -146,7 +150,9 @@ def _react_agent_factory(agent: ReActAgent, *, metadata: ACPServerMetadata | Non
 ACPServer.register_factory(ReActAgent, _react_agent_factory)
 
 
-def _tool_calling_agent_factory(agent: ToolCallingAgent, *, metadata: ACPServerMetadata | None = None) -> ACPAgent:
+def _tool_calling_agent_factory(
+    agent: ToolCallingAgent, *, metadata: ACPServerMetadata | None = None
+) -> ACPServerAgent:
     if metadata is None:
         metadata = {}
 
@@ -173,7 +179,7 @@ def _tool_calling_agent_factory(agent: ToolCallingAgent, *, metadata: ACPServerM
             if isinstance(data, ToolCallingAgentSuccessEvent) and data.state.result is not None:
                 yield acp_models.MessagePart(content=data.state.result.text, role="assistant")  # type: ignore[call-arg]
 
-    return ACPAgent(
+    return ACPServerAgent(
         fn=run,
         name=metadata.get("name", agent.meta.name),
         description=metadata.get("description", agent.meta.description),
@@ -182,3 +188,60 @@ def _tool_calling_agent_factory(agent: ToolCallingAgent, *, metadata: ACPServerM
 
 
 ACPServer.register_factory(ToolCallingAgent, _tool_calling_agent_factory)
+
+
+class ACPServerConfig(BaseModel):
+    """Configuration for the ACPServer."""
+
+    configure_logger: bool | None = None
+    configure_telemetry: bool | None = None
+    self_registration: bool | None = False
+    run_limit: int | None = None
+    run_ttl: timedelta | None = None
+    host: str | None = None
+    port: int | None = None
+    uds: str | None = None
+    fd: int | None = None
+    loop: uvicorn.config.LoopSetupType | None = None
+    http: type[asyncio.Protocol] | uvicorn.config.HTTPProtocolType | None = None
+    ws: type[asyncio.Protocol] | uvicorn.config.WSProtocolType | None = None
+    ws_max_size: int | None = None
+    ws_max_queue: int | None = None
+    ws_ping_interval: float | None = None
+    ws_ping_timeout: float | None = None
+    ws_per_message_deflate: bool | None = None
+    lifespan: uvicorn.config.LifespanType | None = None
+    env_file: str | os.PathLike[str] | None = None
+    log_config: dict[str, Any] | str | None = None
+    log_level: str | int | None = None
+    access_log: bool | None = None
+    use_colors: bool | None = None
+    interface: uvicorn.config.InterfaceType | None = None
+    reload: bool | None = None
+    reload_dirs: list[str] | str | None = None
+    reload_delay: float | None = None
+    reload_includes: list[str] | str | None = None
+    reload_excludes: list[str] | str | None = None
+    workers: int | None = None
+    proxy_headers: bool | None = None
+    server_header: bool | None = None
+    date_header: bool | None = None
+    forwarded_allow_ips: list[str] | str | None = None
+    root_path: str | None = None
+    limit_concurrency: int | None = None
+    limit_max_requests: int | None = None
+    backlog: int | None = None
+    timeout_keep_alive: int | None = None
+    timeout_notify: int | None = None
+    timeout_graceful_shutdown: int | None = None
+    callback_notify: Callable[..., Awaitable[None]] | None = None
+    ssl_keyfile: str | os.PathLike[str] | None = None
+    ssl_certfile: str | os.PathLike[str] | None = None
+    ssl_keyfile_password: str | None = None
+    ssl_version: int | None = None
+    ssl_cert_reqs: int | None = None
+    ssl_ca_certs: str | None = None
+    ssl_ciphers: str | None = None
+    headers: list[tuple[str, str]] | None = None
+    factory: bool | None = None
+    h11_max_incomplete_event_size: int | None = None
